Change \not glyph to non-combining character
Move `\not` glyph (`\@not`) from U+338 (COMBINING LONG SOLIDUS OVERLAY)
to U+2044 (FRACTION SLASH).  Both choices are a hack; what we want is a
non-combining LONG SOLIDUS, but that doesn't exist in Unicode.
FRACTION SLASH seems unlikely to be used by a user, as it's not a
mathematical operator (DIVISION SLASH is for that), and even if they
did, this symbol is not a terrible rendering of a slash.

This aims to fix https://github.com/KaTeX/KaTeX/issues/1842
which reveals that combining characters are problematic on Chrome 72.